requirements.txt
APScheduler==3.6.0
arrow==0.13.1
asn1crypto==0.24.0
astroid==2.3.2
async-timeout==3.0.1
Babel==2.6.0
DateTime==4.3
Django==2.2.6
django-allauth==0.40.0
django-cors-headers==3.1.1
django-environ==0.4.5
django-filter==2.1.0
django-rest-auth==0.9.5
django-slack==5.13.0
django-webpack-loader==0.6.0
djangorestframework==3.9.2
djongo==1.2.36
dnspython==1.16.0
dnspython3==1.15.0
docutils==0.14
entrypoints==0.3
flatten-json==0.1.7
gcloud==0.17.0
geographiclib==1.49
geopy==1.18.1
gevent==1.4.0
google-api-core==1.9.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==0.29.1
google-cloud-firestore==0.32.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
graphviz==0.13
greenlet==0.4.15
grpcio==1.20.0
gunicorn==19.9.0
h5py==2.9.0
helpers==0.2.0
httplib2==0.12.3
hyperlink==19.0.0
idna==2.8
imagesize==1.1.0
incremental==17.5.0
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jws==0.1.3
lazy-object-proxy==1.4.2
llvmlite==0.28.0
lxml==4.3.3
mutagen==1.42.0
nbconvert==5.5.0
nbformat==4.4.0
nose==1.3.7
PyDispatcher==2.0.5
Pygments==2.3.1
PyHamcrest==1.9.0
pylint==2.4.3
pymongo==3.9.0
pyod==0.7.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
PyQt5==5.12.1
PyQt5-sip==4.19.15
pyrsistent==0.15.2
PySide2==5.12.1
python-dateutil==2.8.0
python-http-client==3.2.1
python-jwt==2.0.1
python3-openid==3.1.0
pytz==2018.9
pyzmq==18.0.1
queuelib==1.5.0
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.7.0
rsa==3.4.2
sendgrid==6.1.0
six==1.12.0
typed-ast==1.4.0
tzlocal==1.5.1
uritemplate==3.0.0
vine==1.3.0
virtualenv==16.6.0
w3lib==1.20.0
watchtower==0.5.5
wcwidth==0.1.7
webencodings==0.5.1
websocket==0.2.1
websocket-client==0.56.0
wrapt==1.11.2
whitenoise==4.1.4
twilio==6.33.1
